finish lecture
