+++ 03_information_flow/assets/export/index.html
@@ -74,9 +74,9 @@
         <section >
             
             <h2>React Information Flow</h2>
-<img src="https://res.cloudinary.com/dlzuobe8h/image/upload/v1663777860/phase2/logo_rato6n.png">
-<p><a href="https://labs.tadigital.com/index.php/2020/03/31/unidirectional-data-flow-in-react/">https://labs.tadigital.com/index.php/2020/03/31/unidirectional-data-flow-in-react/</a></p>
-
+<!-- <img src="https://res.cloudinary.com/dlzuobe8h/image/upload/v1663777860/phase2/logo_rato6n.png">
+<p><a href="https://labs.tadigital.com/index.php/2020/03/31/unidirectional-data-flow-in-react/">https://labs.tadigital.com/index.php/2020/03/31/unidirectional-data-flow-in-react/</a></p> -->
+            <img src="../information-flow-diagram.png">
             </section>
