+++ 03_information_flow/project_showcase/src/App.js
@@ -1,14 +1,16 @@
+import { useState } from "react";
 import Header from "./components/navigation/Header";
-import ProjectForm from "./components/project/ProjectForm";
-import ProjectList from "./components/project/ProjectList";
+import ProjectsContainer from "./components/project/ProjectsContainer";
 
 const App = () => {
+  const [isDarkMode, setIsDarkMode] = useState(true);
+
+  const toggleDarkMode = () => setIsDarkMode(!isDarkMode)
 
   return (
-    <div className="App">
-      <Header />
-      <ProjectForm />
-      <ProjectList/>
+    <div className={isDarkMode ? "App" : "App light"}>
+      <Header isDarkMode={isDarkMode} toggleDarkMode={toggleDarkMode} />
+      <ProjectsContainer />
     </div>
   );
 };
