+++ 03_information_flow/project_showcase/src/components/project/SearchBar.js
@@ -1,0 +1,15 @@
+const SearchBar = ({setSearchQuery}) => {
+    const handleSearch = (e) => {
+        setSearchQuery(e.target.value)
+    }
+
+    return (
+        <input
+            type="text"
+            placeholder="Search..."
+            onChange={handleSearch}
+        />
+    );
+}
+
+export default SearchBar;
